Add O6U class to interact with o6u.edu.eg website (WIP)
This class provides methods for initializing a new browser instance and opening new pages on the 'o6u.edu.eg' website.
It also includes a method for closing the browser.

NOTE: This class is a work in progress and is not yet finished.

This will be useful for automating tasks on the o6u.edu.eg website, such as logging in, filling out forms, and extracting data.